get it done
OPL UNFINISHED

+++ src/engine/platform/opl.h
@@ -98,11 +98,11 @@ class DivPlatformOPL: public DivDispatch {
     DivDispatchOscBuffer* oscBuf[44];
     bool isMuted[44];
     struct QueuedWrite {
-      unsigned short addr;
+      unsigned int addr;
       unsigned char val;
       bool addrOrVal;
       QueuedWrite(): addr(0), val(0), addrOrVal(false) {}
-      QueuedWrite(unsigned short a, unsigned char v): addr(a), val(v), addrOrVal(false) {}
+      QueuedWrite(unsigned int a, unsigned char v): addr(a), val(v), addrOrVal(false) {}
     };
     FixedQueue<QueuedWrite,4096> writes;
